Site rendering moved to AppContent
