+++ src/App.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import ListSpaces from './app/ListSpaces';
+import AppContent from './app/AppContent';
 import Box from '@material-ui/core/Box';
 
 function App() {
-  return (
-	<Box display="flex" justifyContent="center" alignItems="center">
-		<Box>
-			<ListSpaces parentSpaceId={-1}></ListSpaces>
+	return (
+		<Box display="flex" justifyContent="center" alignItems="center">
+			<Box>
+				<AppContent parentSpaceId={-1}></AppContent>
+			</Box>
 		</Box>
-	</Box>
-  );
+	);
 }
 
 export default App;
