+++ src/app/GetDevices.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default class GetDevices {
+
+}
