+++ src/app/ListSpacesStyles.tsx
@@ -1,0 +1,8 @@
+import { makeStyles } from "@material-ui/core";
+
+export const useStyles = makeStyles((theme) => ({
+    root: {
+        margin: 10,
+        padding: 10
+    }
+}));
